Remove spoken embedding search; transcript search is keyword-only
NLContextualEmbedding is a fine-tuning backbone, not a retrieval
embedder: mean-pooled token vectors are anisotropic and uncalibrated,
so semantic spoken search produced false positives no threshold tuning
could fix. Drop the whole embedding tier and keep the reliable keyword
matcher.

- Delete SpokenEmbedder, SpokenModel, SpokenIndexer (window+embed).
- Strip SpokenSearch to its keyword tier, rename TranscriptSearch, move
  to Transcription/ beside the cache it reads. Search/ is now a clean
  visual-only (SigLIP) subsystem; the visual/transcript blend already
  lives in the UI/agent call sites, not the coordinator.
- Drop now-dead EmbeddingStore.Header.sourceStamp and
  TranscriptCache.diskStamp (existed only to rebuild spoken indexes).
- Coordinator still transcribes on index; only the post-transcript
  embedding pass is gone.
