+++ Sources/PalmierPro/Search/Indexing/EmbeddingStore.swift
@@ -10,7 +10,6 @@ struct EmbeddingStore {
         let samplerVersion: Int
         let dim: Int
         let count: Int
-        var sourceStamp: String?
     }
 
     struct Row {
