+++ Sources/PalmierPro/Search/SearchIndexCoordinator.swift
@@ -94,8 +94,7 @@ final class SearchIndexCoordinator {
         guard !queue.contains(asset.id), !failedIds.contains(asset.id) else { return }
         let needsVisual = (asset.type == .video || asset.type == .image)
             && VisualIndexer.needsIndex(url: asset.url, spec: model.spec)
-        let needsSpoken = Self.wantsTranscript(asset) && SpokenIndexer.needsIndex(url: asset.url)
-        guard needsVisual || needsSpoken || needsTranscript(asset) else { return }
+        guard needsVisual || needsTranscript(asset) else { return }
         queue.append(asset.id)
         batchTotal += 1
         ensureWorker()
@@ -173,7 +172,6 @@ final class SearchIndexCoordinator {
                     try await SearchIndexCoordinator.waitWhileExportActive()
                     _ = try await TranscriptCache.shared.transcript(for: url, isVideo: isVideo, range: nil)
                 }
-                try await SpokenIndexer.index(url: url)
             }()
             switch asset.type {
             case .image:
